+++ src/model_config/graphic.py
@@ -1,0 +1,19 @@
+"""
+Node Notes
+==========
+
+This is where each graphics comes in play for each of the nodes
+"""
+
+
+class TextFieldInput:
+	pass
+
+
+class DropDownMenu:
+	pass
+
+
+class CheckBox:
+	pass
+