+++ travis/test.py
@@ -1,2 +1,3 @@
 print("Hello World")
 assert 1 == 1
+assert 45 == 0